5/5 23:48 测试不能~~~

+++ ConsoleApplication1/Field.cpp
@@ -1,6 +1,6 @@
 #include "Field.h"
 
-Field::Field(const double& creationTime, const Point& s_Position,const double& s_Amplitude, const double& s_Frequency, const double& s_InitialPhase, const int& s_Speed)
+Field::Field(const double& creationTime, const Point& s_Position,const double& s_Amplitude, const double& s_Frequency, const double& s_InitialPhase, const double& s_Speed)
     : CreationTime(creationTime),SourcePosition(s_Position), SourceAmplitude(s_Amplitude), SourceFrequency(s_Frequency), SourceInitialPhase(s_InitialPhase), SourceSpeed(s_Speed)
 {
     //初始化MediumLayer：MediumLayer中每个Medium接收Field的创建时间
